Move ISR to its own file
Move everything from ATMlib.h to atm_synth.h except the ATMsynth class declaration.
